actividad matricula-alumno-asignatura

+++ ejercicios/matriculas/src/main/java/ies/jandula/matriculas/entities/CursoEtapaId.java
@@ -13,7 +13,7 @@ public class CursoEtapaId implements Serializable
 {
 	private static final long serialVersionUID = 1L;
 
-	private long curso;
+	private int curso;
 	
 	private String etapa;
